Update mainnet contracts

+++ contracts/interfaces/uniswap/IUniswapV2Factory.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.12;
 
 interface IUniswapV2Factory {
-	event PairCreated(address indexed token0, address indexed token1, address pair, uint);
+	event PairCreated(address indexed token0, address indexed token1, address pair, uint256);
 
 	function feeTo() external view returns (address);
 
@@ -10,9 +10,9 @@ interface IUniswapV2Factory {
 
 	function getPair(address tokenA, address tokenB) external view returns (address pair);
 
-	function allPairs(uint) external view returns (address pair);
+	function allPairs(uint256) external view returns (address pair);
 
-	function allPairsLength() external view returns (uint);
+	function allPairsLength() external view returns (uint256);
 
 	function createPair(address tokenA, address tokenB) external returns (address pair);
